// lib/schema/types.ts
export interface ImageObject {
  '@type': 'ImageObject';
  url: string;
  width?: number;
  height?: number;
  caption?: string;
}
// ...
export interface Brand {
  '@type': 'Brand';
  name: string;
  url?: string;
  logo?: string | ImageObject;
  description?: string;
}